Fix initial swipe direction and split app helpers
