+++ src/app.js
@@ -2,6 +2,21 @@ import {createInitialState} from './state.js';
 import {renderSlideContent} from './parser.js';
 import {AudioController} from './audio.js';
 import {loadDeckFromDirectory, loadDeckFromZip, loadDeckFromRemote} from './loaders.js';
+import {
+    canPlayInAudioElement,
+    inferAudioType,
+    isPlayableAudioType,
+    isTextAudioType,
+    preserveStructuredText,
+    ssmlToDisplayText,
+} from './transcript.js';
+import {
+    createSwipeState,
+    finishSwipe,
+    resetSwipeState,
+    startSwipeTracking,
+    updateSwipeTracking,
+} from './swipe.js';
 
 const state = createInitialState();
 let slideAdvanceTimer = null;
@@ -11,17 +26,9 @@ const SLIDE_CHANGE_BELL_STRIKE_SECONDS = 0.025;
 const SLIDE_CHANGE_BELL_DECAY_SECONDS = 3.2;
 const SLIDE_CHANGE_BELL_PAUSE_SECONDS = 0.7;
 const TRANSITION_UNLOCK_TIMEOUT_MS = 10_000;
-const SWIPE_MIN_HORIZONTAL_DISTANCE_PX = 45;
-const SWIPE_MAX_VERTICAL_DRIFT_PX = 120;
 let isSlideTransitionInProgress = false;
 let transitionUnlockTimer = null;
-const swipeState = {
-    startX: 0,
-    startY: 0,
-    lastX: 0,
-    lastY: 0,
-    tracking: false,
-};
+const swipeState = createSwipeState();
 
 const elements = {
     deckTitle: document.querySelector('#deck-title'),
@@ -161,24 +168,22 @@ function bindEvents() {
     elements.slideStage.addEventListener('touchend', (event) => {
         void handleTouchEnd(event);
     }, {passive: true});
-    elements.slideStage.addEventListener('touchcancel', resetSwipeState, {passive: true});
+    elements.slideStage.addEventListener('touchcancel', () => {
+        resetSwipeState(swipeState);
+    }, {passive: true});
 }
 
 function handleTouchStart(event) {
     if (event.touches.length !== 1) {
-        resetSwipeState();
+        resetSwipeState(swipeState);
         return;
     }
     if (event.target instanceof HTMLInputElement || event.target instanceof HTMLButtonElement) {
-        resetSwipeState();
+        resetSwipeState(swipeState);
         return;
     }
     const touch = event.touches[0];
-    swipeState.startX = touch.clientX;
-    swipeState.startY = touch.clientY;
-    swipeState.lastX = touch.clientX;
-    swipeState.lastY = touch.clientY;
-    swipeState.tracking = true;
+    startSwipeTracking(swipeState, touch);
 }
 
 function handleTouchMove(event) {
@@ -186,45 +191,29 @@ function handleTouchMove(event) {
         return;
     }
     const touch = event.touches[0];
-    swipeState.lastX = touch.clientX;
-    swipeState.lastY = touch.clientY;
+    updateSwipeTracking(swipeState, touch);
 }
 
 async function handleTouchEnd(event) {
     if (!swipeState.tracking || !state.deck || isSlideTransitionInProgress) {
-        resetSwipeState();
+        resetSwipeState(swipeState);
         return;
     }
 
-    const changedTouch = event.changedTouches?.[0];
-    const endX = changedTouch?.clientX ?? swipeState.lastX;
-    const endY = changedTouch?.clientY ?? swipeState.lastY;
-    const horizontalDistance = endX - swipeState.startX;
-    const verticalDistance = endY - swipeState.startY;
-    const isHorizontalSwipe = Math.abs(horizontalDistance) >= SWIPE_MIN_HORIZONTAL_DISTANCE_PX
-        && Math.abs(verticalDistance) <= SWIPE_MAX_VERTICAL_DRIFT_PX
-        && Math.abs(horizontalDistance) > Math.abs(verticalDistance);
+    const {isHorizontalSwipe, horizontalDistance} = finishSwipe(swipeState, event.changedTouches?.[0]);
 
-    resetSwipeState();
+    resetSwipeState(swipeState);
 
     if (!isHorizontalSwipe) {
         return;
     }
 
     if (horizontalDistance < 0) {
-        await goToSlide(state.currentIndex - 1);
+        await goToSlide(state.currentIndex + 1);
         return;
     }
 
-    await goToSlide(state.currentIndex + 1);
-}
-
-function resetSwipeState() {
-    swipeState.startX = 0;
-    swipeState.startY = 0;
-    swipeState.lastX = 0;
-    swipeState.lastY = 0;
-    swipeState.tracking = false;
+    await goToSlide(state.currentIndex - 1);
 }
 
 function clearSlideAdvanceTimer() {
@@ -751,123 +740,4 @@ function escapeHtml(value) {
         .replaceAll('>', '&gt;')
         .replaceAll('"', '&quot;')
         .replaceAll("'", '&#39;');
-}
-
-function inferAudioType(audioConfig = {}) {
-    const explicitType = String(audioConfig.type || '').trim().toLowerCase();
-    if (explicitType) {
-        return explicitType;
-    }
-
-    const src = String(audioConfig.src || '').trim();
-    if (!src) {
-        return '';
-    }
-
-    return src
-        .split('#', 1)[0]
-        .split('?', 1)[0]
-        .split('.')
-        .pop()
-        ?.toLowerCase();
-}
-
-function ssmlToDisplayText(ssml) {
-    const normalized = String(ssml).replace(/\r\n?/g, '\n');
-    const withStructure = normalized
-        .replace(/<\s*break\b[^>]*\/?>/gi, '\n')
-        .replace(/<\s*br\s*\/?>/gi, '\n')
-        .replace(/<\s*\/\s*(p|div|section|article|li|ul|ol|h1|h2|h3|h4|h5|h6)\s*>/gi, '\n')
-        .replace(/<\s*\/\s*(s|sentence)\s*>/gi, '\n')
-        .replace(/<\s*\/\s*(speak|paragraph)\s*>/gi, '\n\n');
-
-    const withoutTags = withStructure.replace(/<[^>]+>/g, '');
-    const decoded = decodeSimpleXmlEntities(withoutTags);
-
-    return preserveStructuredText(decoded);
-}
-
-function isTextAudioType(audioType) {
-    return audioType === 'txt' || audioType === 'ssml';
-}
-
-function isPlayableAudioType(audioType, sourcePath = '') {
-    if (audioType && !isTextAudioType(audioType)) {
-        return true;
-    }
-
-    const extension = String(sourcePath)
-        .split('#', 1)[0]
-        .split('?', 1)[0]
-        .split('.')
-        .pop()
-        ?.toLowerCase();
-
-    return ['mp3', 'm4a', 'aac', 'wav', 'ogg', 'oga', 'opus', 'flac', 'webm', 'mp4'].includes(extension || '');
-}
-
-function preserveStructuredText(text) {
-    return String(text)
-        .replace(/\r\n?/g, '\n')
-        .replace(/[ \t]+\n/g, '\n')
-        .replace(/\n{3,}/g, '\n\n')
-        .trim();
-}
-
-function decodeSimpleXmlEntities(text) {
-    return String(text)
-        .replaceAll('&nbsp;', ' ')
-        .replaceAll('&amp;', '&')
-        .replaceAll('&lt;', '<')
-        .replaceAll('&gt;', '>')
-        .replaceAll('&quot;', '"')
-        .replaceAll('&apos;', "'");
-}
-
-function canPlayInAudioElement(audioElement, audioType, sourcePath) {
-    const mimeType = inferMimeType(audioType, sourcePath);
-    if (!mimeType) {
-        return false;
-    }
-    return audioElement.canPlayType(mimeType) !== '';
-}
-
-function inferMimeType(audioType, sourcePath) {
-    const normalizedType = String(audioType || '').toLowerCase();
-    if (normalizedType === 'ssml') return 'application/ssml+xml';
-    if (normalizedType === 'txt') return 'text/plain';
-    if (normalizedType === 'mp3') return 'audio/mpeg';
-    if (normalizedType === 'wav') return 'audio/wav';
-    if (normalizedType === 'ogg') return 'audio/ogg';
-    if (normalizedType === 'oga') return 'audio/ogg';
-    if (normalizedType === 'opus') return 'audio/ogg; codecs=opus';
-    if (normalizedType === 'm4a') return 'audio/mp4';
-    if (normalizedType === 'aac') return 'audio/aac';
-    if (normalizedType === 'flac') return 'audio/flac';
-    if (normalizedType === 'webm') return 'audio/webm';
-    if (normalizedType === 'mp4') return 'audio/mp4';
-
-    const extension = String(sourcePath)
-        .split('#', 1)[0]
-        .split('?', 1)[0]
-        .split('.')
-        .pop()
-        ?.toLowerCase();
-
-    const extensionToMime = {
-        ssml: 'application/ssml+xml',
-        txt: 'text/plain',
-        mp3: 'audio/mpeg',
-        wav: 'audio/wav',
-        ogg: 'audio/ogg',
-        oga: 'audio/ogg',
-        opus: 'audio/ogg; codecs=opus',
-        m4a: 'audio/mp4',
-        aac: 'audio/aac',
-        flac: 'audio/flac',
-        webm: 'audio/webm',
-        mp4: 'audio/mp4',
-    };
-
-    return extensionToMime[extension] || '';
 }
